# webpage/platform-integration/docs/openapi-platform-mock.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Platform Integration Mock API",
    "description": "FastAPI 骨架，提供积分、授权、任务与审核回写的 Mock 接口供联调使用。",
    "version": "0.1.0"
  },
  "paths": {
    "/api/v1/credits/estimate": {
      "post": {
        "tags": [
          "credits"
        ],
        "summary": "积分试算",
        "operationId": "estimate_api_v1_credits_estimate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EstimateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EstimateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/credits/charge": {
      "post": {
        "tags": [
          "credits"
        ],
        "summary": "直接扣费",
        "operationId": "charge_api_v1_credits_charge_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChargeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChargeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/credits/ledger": {
      "get": {
        "tags": [
          "credits"
        ],
        "summary": "积分流水",
        "operationId": "ledger_api_v1_credits_ledger_get",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tenant Id"
            }
          },
          {
            "name": "user_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LedgerResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/credits/pre-deduct": {
      "post": {
        "tags": [
          "credits"
        ],
        "summary": "预扣积分",
        "operationId": "pre_deduct_api_v1_credits_pre_deduct_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PreDeductRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreDeductResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/credits/commit": {
      "post": {
        "tags": [
          "credits"
        ],
        "summary": "结算扣费",
        "operationId": "commit_api_v1_credits_commit_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommitRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreDeductResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/credits/cancel": {
      "post": {
        "tags": [
          "credits"
        ],
        "summary": "取消预扣",
        "operationId": "cancel_api_v1_credits_cancel_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CancelRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object",
                  "title": "Response Cancel Api V1 Credits Cancel Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/credits/balance": {
      "get": {
        "tags": [
          "credits"
        ],
        "summary": "查询积分余额",
        "operationId": "balance_api_v1_credits_balance_get",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tenant Id"
            }
          },
          {
            "name": "user_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BalanceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BalanceResponse": {
        "properties": {
          "tenant_id": {
            "type": "string",
            "title": "Tenant Id"
          },
          "user_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Id"
          },
          "balance": {
            "type": "integer",
            "title": "Balance"
          },
          "frozen": {
            "type": "integer",
            "title": "Frozen"
          },
          "currency": {
            "type": "string",
            "title": "Currency",
            "default": "point"
          }
        },
        "type": "object",
        "required": [
          "tenant_id",
          "balance",
          "frozen"
        ],
        "title": "BalanceResponse"
      },
      "CancelRequest": {
        "properties": {
          "pre_deduct_id": {
            "type": "string",
            "title": "Pre Deduct Id"
          },
          "tenant_id": {
            "type": "string",
            "title": "Tenant Id"
          },
          "user_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Id"
          },
          "reason": {
            "type": "string",
            "title": "Reason",
            "description": "取消原因"
          }
        },
        "type": "object",
        "required": [
          "pre_deduct_id",
          "tenant_id",
          "reason"
        ],
        "title": "CancelRequest"
      },
      "ChargeRequest": {
        "properties": {
          "tenant_id": {
            "type": "string",
            "title": "Tenant Id"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "template_id": {
            "type": "string",
            "title": "Template Id"
          },
          "task_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Task Id",
            "description": "关联任务 ID，可为空"
          },
          "amount": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Amount",
            "description": "直接扣费积分"
          },
          "priority": {
            "type": "string",
            "title": "Priority",
            "description": "扣费对应的优先级模式",
            "default": "standard"
          },
          "reason": {
            "type": "string",
            "title": "Reason",
            "description": "扣费原因标签",
            "default": "charge"
          }
        },
        "type": "object",
        "required": [
          "tenant_id",
          "user_id",
          "template_id",
          "amount"
        ],
        "title": "ChargeRequest"
      },
      "ChargeResponse": {
        "properties": {
          "ledger_id": {
            "type": "string",
            "title": "Ledger Id"
          },
          "tenant_id": {
            "type": "string",
            "title": "Tenant Id"
          },
          "user_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Id"
          },
          "task_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Task Id"
          },
          "balance_after": {
            "type": "integer",
            "title": "Balance After"
          },
          "change": {
            "type": "integer",
            "title": "Change"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "ledger_id",
          "tenant_id",
          "user_id",
          "task_id",
          "balance_after",
          "change",
          "reason",
          "created_at"
        ],
        "title": "ChargeResponse"
      },
      "CommitRequest": {
        "properties": {
          "pre_deduct_id": {
            "type": "string",
            "title": "Pre Deduct Id"
          },
          "actual_cost": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Actual Cost"
          },
          "task_id": {
            "type": "string",
            "title": "Task Id"
          },
          "tenant_id": {
            "type": "string",
            "title": "Tenant Id"
          },
          "template_id": {
            "type": "string",
            "title": "Template Id"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          }
        },
        "type": "object",
        "required": [
          "pre_deduct_id",
          "actual_cost",
          "task_id",
          "tenant_id",
          "template_id",
          "user_id"
        ],
        "title": "CommitRequest"
      },
      "EstimateOption": {
        "properties": {
          "priority": {
            "type": "string",
            "title": "Priority"
          },
          "total_cost": {
            "type": "integer",
            "title": "Total Cost"
          },
          "eta_minutes": {
            "type": "integer",
            "title": "Eta Minutes"
          }
        },
        "type": "object",
        "required": [
          "priority",
          "total_cost",
          "eta_minutes"
        ],
        "title": "EstimateOption"
      },
      "EstimateRequest": {
        "properties": {
          "tenant_id": {
            "type": "string",
            "title": "Tenant Id"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "template_id": {
            "type": "string",
            "title": "Template Id"
          },
          "scene": {
            "type": "string",
            "title": "Scene"
          },
          "resolution": {
            "type": "string",
            "title": "Resolution",
            "description": "分辨率标识：sd/hd/4k",
            "default": "hd"
          },
          "priority": {
            "type": "string",
            "title": "Priority",
            "description": "standard/accelerated",
            "default": "standard"
          },
          "extras": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "integer"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extras",
            "description": "附加消耗，如超分/高清等"
          }
        },
        "type": "object",
        "required": [
          "tenant_id",
          "user_id",
          "template_id",
          "scene"
        ],
        "title": "EstimateRequest"
      },
      "EstimateResponse": {
        "properties": {
          "currency": {
            "type": "string",
            "title": "Currency"
          },
          "template_id": {
            "type": "string",
            "title": "Template Id"
          },
          "scene": {
            "type": "string",
            "title": "Scene"
          },
          "resolution": {
            "type": "string",
            "title": "Resolution"
          },
          "priority": {
            "type": "string",
            "title": "Priority"
          },
          "current_balance": {
            "type": "integer",
            "title": "Current Balance"
          },
          "selected_cost": {
            "type": "integer",
            "title": "Selected Cost"
          },
          "suggest_topup": {
            "type": "integer",
            "title": "Suggest Topup"
          },
          "extras": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Extras"
          },
          "options": {
            "items": {
              "$ref": "#/components/schemas/EstimateOption"
            },
            "type": "array",
            "title": "Options"
          }
        },
        "type": "object",
        "required": [
          "currency",
          "template_id",
          "scene",
          "resolution",
          "priority",
          "current_balance",
          "selected_cost",
          "suggest_topup",
          "extras",
          "options"
        ],
        "title": "EstimateResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LedgerEntryResponse": {
        "properties": {
          "ledger_id": {
            "type": "string",
            "title": "Ledger Id"
          },
          "task_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Task Id"
          },
          "change": {
            "type": "integer",
            "title": "Change"
          },
          "balance_after": {
            "type": "integer",
            "title": "Balance After"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "ledger_id",
          "task_id",
          "change",
          "balance_after",
          "reason",
          "created_at"
        ],
        "title": "LedgerEntryResponse"
      },
      "LedgerResponse": {
        "properties": {
          "tenant_id": {
            "type": "string",
            "title": "Tenant Id"
          },
          "user_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Id"
          },
          "entries": {
            "items": {
              "$ref": "#/components/schemas/LedgerEntryResponse"
            },
            "type": "array",
            "title": "Entries"
          }
        },
        "type": "object",
        "required": [
          "tenant_id",
          "user_id",
          "entries"
        ],
        "title": "LedgerResponse"
      },
      "PreDeductRequest": {
        "properties": {
          "task_id": {
            "type": "string",
            "title": "Task Id",
            "description": "任务唯一 ID"
          },
          "template_id": {
            "type": "string",
            "title": "Template Id",
            "description": "模板 ID"
          },
          "tenant_id": {
            "type": "string",
            "title": "Tenant Id",
            "description": "创作者/租户 ID"
          },
          "user_id": {
            "type": "string",
            "title": "User Id",
            "description": "授权用户 ID"
          },
          "scene": {
            "type": "string",
            "title": "Scene",
            "description": "任务场景标识"
          },
          "estimated_cost": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Estimated Cost",
            "description": "预估积分消耗"
          },
          "currency": {
            "type": "string",
            "title": "Currency",
            "description": "积分计价单位",
            "default": "point"
          },
          "expire_in": {
            "type": "integer",
            "maximum": 3600.0,
            "minimum": 60.0,
            "title": "Expire In",
            "description": "预扣单过期秒数",
            "default": 600
          }
        },
        "type": "object",
        "required": [
          "task_id",
          "template_id",
          "tenant_id",
          "user_id",
          "scene",
          "estimated_cost"
        ],
        "title": "PreDeductRequest"
      },
      "PreDeductResponse": {
        "properties": {
          "pre_deduct_id": {
            "type": "string",
            "title": "Pre Deduct Id"
          },
          "frozen_amount": {
            "type": "integer",
            "title": "Frozen Amount"
          },
          "balance_after": {
            "type": "integer",
            "title": "Balance After"
          },
          "quota_snapshot": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Quota Snapshot"
          },
          "expire_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expire At"
          }
        },
        "type": "object",
        "required": [
          "pre_deduct_id",
          "frozen_amount",
          "balance_after",
          "quota_snapshot",
          "expire_at"
        ],
        "title": "PreDeductResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}